- fixed some warnings

+++ source/duke3d/src/game.cpp
@@ -3252,7 +3252,7 @@ int A_Spawn(int spriteNum, int tileNum)
                     int const startWall = sector[sectNum].wallptr;
                     int const endWall   = startWall + sector[sectNum].wallnum;
 
-                    pSprite->extra = ((uint16_t)sector[sectNum].hitag != UINT16_MAX);
+                    pSprite->extra = (uint16_t)((uint16_t)sector[sectNum].hitag != UINT16_MAX);
 
                     // TRAIN_SECTOR_TO_SE_INDEX
                     sector[sectNum].hitag = newSprite;
